Move InterestGroup and auctionEvents to common package
